Quick fixes

+++ racional/src/racional.h
@@ -5,11 +5,11 @@
 
 class Racional{
 private:
-    double num_, denom_;
+    int num_, denom_;
 public:
-    Racional(double arg1 = 0, double arg2 = 0);
-    double getNum();
-    double getDenom();
+    Racional(int arg1 = 0, int arg2 = 0);
+    int getNum();
+    int getDenom();
     friend bool operator==(const Racional &obj1, const Racional &rac);
     friend std::istream& operator>>(std::istream &stream_in, Racional &rac);
     friend std::ostream& operator<<(std::ostream &stream_out, const Racional &rac);
